remove Tunapplied_alias constructor
Summary:
This diff is part of a series that removes the implementation of higher-kinded types from the type-checker.

This diff removes the constructor `Tunapplied_alias` from `Typing_defs_core.ty_`. This constructor was only ever used for HK types:

Given
```
type Foo<T> = ...
```
we used `Tunapplied_alias "Foo"` to represent the localized version of the higher-kinded type `Foo` (i.e., without type arguments). This was just a workaround for the fact that type alias are otherwise just substituted away during localization.

Reviewed By: vassilmladenov

Differential Revision: D73776221

fbshipit-source-id: 5ed77ddd48bf70ed88fa6b276dc8367ebe11cf66