remove type args from Tgeneric
Summary:
This diff is part of a series that removes the implementation of higher-kinded types from the type-checker.

This diff removes the list of type arguments from `Tgeneric` in `Typing_defs_core.ty_`. These arguments were only used if the type parameter itself was higher-kinded. Therefore, these are no longer used.

Since this diff touches almost every usage of `Tgeneric`, I've kept all changes as mechanical as possible.

In particular, there are lots of patterns of the form
```
Tgeneric(name, tyargs) -> ...
```

which I replaced with
```
Tgeneric(name) ->
  let tyargs = [] in
  ...
```
if the required changes are non-trivial.

I've left TODOs in these cases to clean this up further.

Reviewed By: viratyosin

Differential Revision: D73855688

fbshipit-source-id: a1bf92e0dbb6a6847b86a6ea9dc549f7f53de3e7