remove type argument param from Env.get_*_bound
Summary:
This diff is part of a series that removes the implementation of higher-kinded types from the type-checker.

In general, in the presence of higher-kinded types, a type variable/`Tgeneric` `T` can have a list of type arguments. When querying the environment for the bounds on  `T`, we need to take these into account.

This is why `Env.get_{lower,upper,equal}_bounds` and related functions currently have a`locl_ty list` parameter, representing the type arguments to the variable.

As of D73855688, the `locl_ty` list passed to `Env.get_{lower,upper,equal}_bounds` is empty, this diff removes that parameter from these functions altogether.

Reviewed By: viratyosin

Differential Revision: D73991990

fbshipit-source-id: ac271970d473da4094f682e08ac95302ec061f6a