ClassTag of id * locl_ty list
Summary:
This diff updates the `ClassTag` variant to include a list of locl_tys in order to represent class predicates with (fully reified) generics.

This diff doesn't actually start creating this kind of predicate, but instead just updates existing code to handle the new possibility of specified arguments. It maintains the implicit convention of "no arguments" for a generic class meaning passing any instance of that class regardless of the generics.

Reviewed By: dlreeves

Differential Revision: D71941112

fbshipit-source-id: 65e6b2d0c9395d4fd769788731a62d72b10f756a