Remove `decl_ty` and `locl_ty` aliases from set of recursive types defining `ty`
Summary:
Our definition of types is a GADT with an type used to index the 'phase'. It also defines type aliases `decl_ty` and `locl_ty` for `decl_phase ty` and `locl_phase ty` respectively. Because these aliases are included in the set of mutually recursive definitions, ppx_transform was generating transformation fields `on_ty_decl_ty` and `on_ty_locl_ty` which give the impression you can define a transformation for the relevant subset of types.

Unfortunately, this impression is misleading - these functions won't be called unless you use the `transform_ty_decl_ty` or `transform_ty_locl_ty` entry points and even then they won't be used in recursive cases - the `on_ty_ty` handler is the one that is called for any nested type.

This is a nasty footgun so this diff moves the aliases out of the recusive set of bindings and rewrites functions to use `on_ty_ty` where possible or introduce manual traversals where we need knowledge of the indexing type and the universally quantified type would escape its scope

Reviewed By: viratyosin

Differential Revision: D74066341

fbshipit-source-id: 93b0b7f66dec063149661477a1100210175c85f9