Rename "DynamicStructPatch" to "<StructPatch>"
Summary:
Most users shouldn't care whether it's a dynamic patch or not. The dynamic patch is more likey an implementation detail -- currently static patch is also implemented by converting to dynamic patch first.

This diff also added `<` and `>` to make it clear it's a special node (similar to `<list>`, `<set>`, etc)

Reviewed By: thedavekwon

Differential Revision: D74108216

fbshipit-source-id: ae95dba45982705136a8190b70cc29c0ba96b957