Avoid flattening the same constant multiple times in HHBBC
Summary:
We might copy a constant onto a class' constant list (if the constant
is abstract with a default) before trait flattening actually
occurs. If, during trait flattening, another trait defines the same
constant, we might flatten *that* as well onto the class' constant
list. This results in two constants with the same name being on the
list, causing all sorts of downstream issues. The fix is to remove
that constant from the "trait cns" list of that class.

Reviewed By: mdko

Differential Revision: D74152572

fbshipit-source-id: f76db29016012d5bc93f6ffd2dfb50c1127c0690