Add handshake logging to transport info for all connections.
Summary:
Previously, we only added this to transport info on the fallback path,
requiring different methods for accessing this info on fizz connections.

This could come at a small memory cost since this adds another copy, however
the TransportInfo struct is not particulalry memory efficient to begin with,
and this will allow gradually removing the rest of the SSL fields, likely at a
memory improvement.

Differential Revision: D74079278

fbshipit-source-id: 64724cf019fab963371ea366907c91b064d2bea1