Shrink ElementWrapper to just the pointer and other optimizations.
Summary:
The per thread ElementWrapper array can be very sparse and for a
process where num_threads * num_thread_local_objects is large, it can produce
significant memory overheads. By moving all the auxilliary information from
ElementWrapper into the more compact ThreadEntrySet, we can eliminate 8 bytes
per wrapper (sizeof(ElementWrapper) goes from 16 bytes to 8).

This also fixes a racing behavior bug in destorying a TL object and one of the
threads that had an instance of it exiting.

As an added bonus, we can also now have accessAllThreads() use the copy of the
per thread instance ptr stashed in ThreadEntrySet instead of accessing the
elements array of a thread directly. The lock protecting the ThreadEntrySet is
already held in the accessor and we no longer need it to also hold the
StaticMeta's lock_. The meta.lock_ was otherwise held to prevent a thread's
element array from being resized when the accessor might be trying to read from
it.

Reviewed By: yfeldblum

Differential Revision: D73058743

fbshipit-source-id: 53786848b369220756ea74dc315056d3ca4f2d2c