reduce direct usages of Watchman.clock type
Summary:
This is a preparation step for adding support for Eden notifications as a backend in `ServerNotifier`, our abstraction over file watching in the server.

Currently, many modules interacting with `ServerNotifier` pass around `Watchman.clock` values. This diff changes those places to use a new type, `ServerNotifierTypes.clock` instead. Note that this isn't replacing all usages of `Watchman.clock` values in Hack, but only those directly interacting with `ServerNotifier`.

The new `ServerNotifierTypes.clock`  type had to go into its own library in order to avoid cyclic dependencies. It's currently a variant type with only on constructor, but that will change in the future.

Differential Revision: D73188033

fbshipit-source-id: 4cfb3257dcc684930aed8547100112110cd99444