tighten ServerNotifier API
Summary:
This diff makes two changes to the API of ServerNotifier:
1. It changes the return type of `get_changes_sync` from `changes` to `SSet.t`.
2. It removes the `StateEnter` and `StateLeave` variants from the `changes` type, which is now only returned by `get_changes_async`.

The first change is motivated by the fact that `get_changes_sync` was already only returning the `SyncChanges` variant of `changes`, and this fact was reflected in a comment on it.

The motivation for the second change is to make the API less Watchman-specific and stop exposing information that is potentially difficult to migrate to Eden as a file watching service.
Currently, the   `StateEnter` and `StateLeave` variants are Watchman-specific and indicate that a Watchman state (such as `hg.update` or `meerkat-build`) was entered or left. However, looking at the call sites of `get_changes_async` we may notice that callers don't actually care about the state transitions, they only use `State_enter`/`State_leave` as part of their logic to determine whether the information they got may be stale or not. However, the `changes` type of `get_changes_async` already allows us to indicate whether the returned result is stale or not (by choosing between `SyncChanges` and `AsyncChanges`).

Differential Revision: D74252978

fbshipit-source-id: 018d4857c8c64c04302c7086cc2f8a5199f21926