Remove TODO for case types with where clauses
Summary:
Expanding along the same conditions as regular case types seems fine.

It seems like expansion is only forced when doing stuff like tasts checks that will work acceptable with a larger type (which is effectively what you get if you union the variants and ignore the where clauses) or when checking for recursion.

Reviewed By: geralt-encore

Differential Revision: D74284141

fbshipit-source-id: 06033ef7cedf027ea275cc57c609a7ddb2935711