provide request during setup
Summary:
In this diff, we save a copy of `CustomCompressionSetupRequest` in rocket client, to facilitate the creation of custom compression.

Previously, we were only able to see the negotiation response when creating a custom compression. With this change, we can see both the negotiation request, and the response. Without having access to the request, we might have to copy of a portion of the request in the response, causing bandwidth waste.

Reviewed By: cevans87

Differential Revision: D73977495

fbshipit-source-id: 26c481ae554b6f4ea12202bb3315b30fad2c0d40