Remove structured key usage in record.thrift
Summary:
This diff removes structured keys in `record.thrift` for Hack and JSON support.

It also means that `SerializableRecord` no longer needs to be *sealed*.

To achieve this, I've removed the adapter and instead move the logic to a layer above. For those looking to maintain the invariants described in `record.thrift`, they should use the `SerializableRecord` class (not `SerializableRecordUnion`). There is some additional cost for marshaling to/from these two representations but in this case, correctness trumps performance.

One upside is that adapter templating non-sense is gone :)

Reviewed By: pranavtbhat

Differential Revision: D74292327

fbshipit-source-id: f6f9fabd6d6c81688e8b2201258e132dda62d7ff