Don't query the host/port from SR unless we need it
Summary:
The current Thrift code _always_ queries the host and port from SR (https://fburl.com/code/9njc2hyu) even though the vast majority of the time this will never be used.  It will only get used for the very few queries that get sampled for logging.

Let's change this up and only retrieve this information when needed for logging.

Also, set the host to "<not set>" on a failure to try to understand why we don't see this information in some samples (https://fburl.com/scuba/mysql_client_logs/81a71rtp).

Differential Revision: D74091091

fbshipit-source-id: 3b78b7547562f40536c963cd8e3db502d1188a76