Add an explanation for assumed references in refcount-opts
Summary: It was a little confusing why we were reducing the lower bound of may-aliases as well here, especially since we only reduce the exclusive supported component of the aset's lower bound. See D19271089 for more context. Thanks paulbiss for pointing to the diff and providing the context behind it.

Reviewed By: jano

Differential Revision: D74020324

fbshipit-source-id: f8c14e2bca5b9102faeac95d9aa7679873d4d55a