op::visit_union_with_tag
Summary:
Combines the ergonomics of visitation's visit_union with the power and nondependence on metadata of always-on reflection.
While `invoke_by_field_id` provides the same capability, the API is not intuitive and there's some boilerplate which this function encapsulates. In addition, the API similarity with `visit_union` will facilitate migration.
`visit_union` is renamed both to help disambiguate and because shorter names may be more appealing than longer ones, while we want to steer new uses to AOR.

Reviewed By: Mizuchi

Differential Revision: D74209943

fbshipit-source-id: 3cc08c173307f9cccdd6a948f52677be53596c8f