codemod program to add @cpp.EnableCustomTypeOrdering
Summary: If a structure with URI is orderable today, and it has set/map field with custom types, it will become unorderable once we disabled ordering custom types. This might break the existing code. To avoid the breakage, we need to add `cpp.EnableCustomTypeOrdering` to those structured to preserve the existing behavior and still makes it orderable.

Reviewed By: aristidisp

Differential Revision: D74455603

fbshipit-source-id: 6675a443a61d03e3501d760967e67b976e1df375