Add get_asserted_states to rust client library
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Diff
Implement the rust client for get_asserted_states. Update example.

# Why
For the migration process, we will be needing to run both edenfs and watchman states in paralell, since some clients will not have migrated all at the same time.

For this, we want to be able to read the watchman state from the edenfs client library. This stack will expose the debug command debug-get-asserted-states via the watchman rust client.

Reviewed By: jdelliot

Differential Revision: D74096311

fbshipit-source-id: 78a2ae744b0738762782c21072dedfa39c0d86f7