Fix non-unique ProfPrologueIDs
Summary:
In low pointer builds we use the 32 bit Func* as an ID instead of computing separate unique ID numbers for each function. This works fine in repo mode but in non-repo mode it means that FuncId is no longer unique (e.g. if I free a Func and allocate a new one it may be allocated at the same address and therefore share the same FuncId).

This is mostly a theoretical concern as we don't run non-repo low pointer builds anywhere but we do have some tests that depend on being able to calculate unique ProfPrologueIDs in all build configurations. To fix that without expanding the Func structure store a side table of IDs in non-repo mode.

Reviewed By: ricklavoie

Differential Revision: D73924655

fbshipit-source-id: 3495f3ab3c8b941f5b95913e2e7dad21923673e3