Refactoring Lazy Two Way Map
Summary:
I want to add unit testing around the lazy two way map, so I'm refactoring to separate that out a bit and to set things up so we don't have to link in half of Facts for the unit test.

This makes LazyTwoWayMap take an additional template argument of the key type, and converts the PathVersions class to a template version provider that uses a function template specialization to help.

Differential Revision: D74533154

fbshipit-source-id: a37740e2020e7a3b9f205100e388171484b78040