edenfs_watcher skeleton
Summary:
This diff contains the initial scaffolding to implement a new EdenFS-based file watching service. It works by calling into Rust and using the `edenfs-client` library there, in particular and its `changes_since` functionally.

The entry point is the `edenfs_watcher` module, which defined the API to be consumed by, say, a new backend for `ServerNotifier` in the future.

The current version is not intended to be production-ready in any way and is not actually called from anywhere. The idea is just to get a lot of the FFI-related boilerplate code (oxidization, etc) out there and reduce the amount of unpublished code I'm carrying around.

Differential Revision: D74322874

fbshipit-source-id: 0261133a8974edea7b723e41dbb9eeaeb845979e