flags for EdenFS file watcher in server config
Summary:
This diff adds a flag to `ServerLocalConfig` to be used in the future to enable EdenFS-backed file watching. The flag isn't actually used.

The new flags follow the convention established for Watchman in D21588418 , meaning that all new flags have the shared prefix `edenfs_file_watcher`.

In particular, this means that the new file watcher is enabled using `edenfs_file_watcher_enabled`, not `use_edenfs_file_watcher`.

I've also added a jk that can override the setting.

Reviewed By: mheiber

Differential Revision: D74478230

fbshipit-source-id: 9c6d0a9e40b6d577d793da30e12fa858c1491029