add Eden (pseudo) backend to ServerNotifier
Summary:
This diff adds a new variant to `ServerNotifier.t`, our main abstraction over what file watching service to use in the server.

The initialization logic in `ServerNotifier` is changed so we use the new variant if the flag `edenfs_file_watcher_enabled` is set. In that case, we initialize a new `Edenfs_watcher.instance`.

This means that there are no changes to existing behavior, unless `edenfs_file_watcher_enabled` is set. Furthermore, even if the flag is set, we don't actually use the `Edenfs_watcher.instance` to provide the `ServerNotifier` functionality. Instead, we *also* create a Watchman instance and forwarding all calls to it. This is due to the fact that `Edenfs_watcher` does not provide the required functionality yet, but I want a working `ServerNotifier` if `edenfs_file_watcher_enabled` is set in order to start creating integration testing in a follow-up diff.

Differential Revision: D74652741

fbshipit-source-id: 2d572f74b9069087acf20bb31f1f41563103dd91