make server process do fbinit
Summary:
This diff changes the initialization of the server process so that instead of initializing Folly, it does `fbinit` instead.

This is required in order for the server process to use EdenFS for file watching in the future. The Eden client library uses thrift, which requires `fbinit`.

To this end, this diff revives the (previously unused) `startup_initializer` library inside the Hack codebase. The Folly initialization is actually replaced, since `fbinit` subsumes it.

Two minor changes:
- In order to be compatible with what `Folly.ensure_folly_init` does with signal handlers, `startup_initializer` is updated  to handle `SIGTERM` the same as `hack/src/facebook/folly/follyClientLib.cpp`
- In order to make sure that we don't accidentally initialize Folly twice, we now call Folly.mark_folly_as_initialized in `startup_initializer`.

Differential Revision: D73036099

fbshipit-source-id: 7fb202ff890462bcdaa78bc7cc2d54539beb60d9