Check that type splat parameter in declaration is a subtype of tuple
Summary:
We weren't checking that a type parameter `T` in a type-splat `...T` is actually a tuple (by subtyping against `(mixed...)`), except when that type-splat appeared nested inside a function-type hint or tuple-type hint.

Note that the errors are not ideal; we repeat essentially the same message in multiple errors. This should be fixed separately.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D74829336

fbshipit-source-id: 240223aa67758a19029143dfb29dea64d382f777