Do not consider class<T> may contain reified
Summary:
This was incorrect and it was breaking the hackc logic that tries to handle the case when an Happly might have a reified argument i.e. not a locally defined generic (reified or not). For the moment, HclassPtr only ever emits an `"HH\class"` type constraint.

For the future, I think we will never have to look, as TReified in `class<TRefied>` is reified but does not contain reified args itself, but added a TODO to follow up when we add the deeper enforcement -- it might make a difference for producing a type structure.

Reviewed By: andrewjkennedy

Differential Revision: D75012205

fbshipit-source-id: b260db54be25b8309ba1b0be8bad59b19dab6162