Index bundled schema data by URI
Summary:
This diff adds the ability to look up a definition node by URI, provided that definition was previously accessed statically to trigger deserializing its bundled schema data.
Stacked diffs will remove the requirement that the definition was previously accessed with dynamic bundling.
We expect to convert the returned node to the equivalent type system node, so the API is private.

Reviewed By: praihan

Differential Revision: D75018690

fbshipit-source-id: ed74ffc94de3de5b9ff1eb1ef63009ceeb511667