Refactor ReifiedGenericsInfo: remove m_hasSoftReified and m_bitmap fields
Summary: As title. These fields can be computed from the information that is present in the m_typeParamInfo field. Both these fields are used when lowering the IsFunReifiedGenericsMatched ir-op. The computation is simple enough that we don't need a field to hold this information. Reclaim a few bytes of memory from ExtendedSharedData.

Reviewed By: jano

Differential Revision: D74902291

fbshipit-source-id: 755d0fcec11d4c8f03ea77b3ade0c84e2688cd07