Further Refactoring of Symbol Map Test
Summary:
Incorporates feedback around removing the VersionKey template argument to LazyTwoWayMap by making the versioning use a StringPtr instead of a Path.  Ultimately, StringPtr is what was the key anyways so this works out nicely.

This also separates the StringData fake implementation used for tests into it's own file.  This will subsequently be used in unit testing for the lazy two way map.

Differential Revision: D75018382

fbshipit-source-id: 21e25da407af901d01f0269f28563e5171f14640