Do not inspect case types
Summary:
If I understand correctly, what this code is doing, it's for optimization that asks "when we fetch A, what else should we pre-fetch because it'll be likely to be needed when A is used."

Unlike simple (non-newtype) type aliases, case types aren't always expanded to their variants--only when the case type appears as a supertype, so it's not necessarily the case that we will read the variants since if it appears in the subtype position, we'll only look at any explicit upper bound.

Also, even before I errantly changed this code to only look at the first variant (in D62251888), this code was usually not looking at case type variants since previously case types were just typedefs with union type assignments and this code didn't look at unions.

Reviewed By: mheiber

Differential Revision: D74282550

fbshipit-source-id: 2e6976c95f8e96bde566d60e3f73020e23918ef2