Permit non <<__Reifiable>> type constants in ReifiedGenerics intrinsic
Summary: In D73391586 I made this a separate error; this removes the check entirely and removes the special code. These intrinsics are meant as replacements for type_structure, which does not do these checks; the input reified types are destroyed by builtins -- so we don't need the input type to be a valid reified witness.

Differential Revision: D75217995

fbshipit-source-id: 4133714487355f3e12a0b6ef39139dacbe4dc9bf