intersect tyvars instead of bounding with intersections
Summary:
When we do type splitting using the constraint type, we create two type variables, one for the true branch and one for the false branch. These variables are bounded below by types and some of these binding types are intersections with the type associated with the predicate or the negative of the predicate.

This diff changes this so that the type variables' bounds just use the left/span/right from the partitioning and we intersect the type associated with the predicate with the type variable itself.

This is a little bit closer to what we were doing originally with refinement (intersection on the outside) and so the idea was that it'd fit nicer with some of the incompleteness in the typechecker with the intersection on the outside--error count wise it's kinda a wash, but:

A more concrete benefit is that when we add support for predicates for classes with erased generics, converting a predicate to a type requires instantiating the generics (kinda like unpacking an existential type). This isn't something we want to do in subtyping, and so by moving where we do the intersection with the type from the predicate, we can keep all the instantiation inside just typing.ml (requires some other changes to other uses of to_ty too)

Reviewed By: dlreeves

Differential Revision: D68849503

fbshipit-source-id: 3120a9ad890df462269da699767485b27e866e04