allow dynamic updates to per key connection pool limits
Summary:
This enables dynamic updates to the per-key connection pool limit.

The Iris team needs the ability to dynamically update this property as a SEV1 followup, since restarting our service in order to increase the connection_cap_per_shard was very slow. We were seeing this error which was affecting our throughput:
```

Reviewed By: jkedgar

Differential Revision: D75182675

fbshipit-source-id: 0683784498e58f365600d9f9f7c063a2f8e3c2c3