determine mount point in Edenfs_watcher
Summary:
Prior to this diff, `Edenfs_watcher` only supports watching a folder if it happens to be the mount point of an Eden mount. In other words, this means that we could only watch the toplevel/root folder of a hg repository mounted with Eden. For example, we can watch `~/www`, but not `~/fbsource/www`, because in the latter the root of the repository is `~/fbsource` and `www` is just a subfolder.

This diff updates the initialization code for `Edenfs_watcher` such that given a `path` to watch, it
- checks if `path` itself is the mount point, or otherwise
- which ancestor of `path` the mount point is and splits off the relative path `root` from the mount point to get the full `path`.

The `root` and mount point are then used when calling the Eden client library.

We can now also detect directly if we are trying to initialize `Edenfs_watcher` on a non-Eden www and produce a dedicated error.

Differential Revision: D75217988

fbshipit-source-id: fcdd344ee8960eb77d2e007ae08c0c338652d712