Name default constants used with TDigest
Summary:
The default values for digest size and buffer size are duplicated in a few places. It makes sense to have them as constants in `TDigest`, as the relationship between the two is somewhat dictated by the relative costs of sorting and merging, so that is a good place to give an authoritative recommendation.

Also clean up a few bits.

Reviewed By: ilvokhin

Differential Revision: D75216992

fbshipit-source-id: cc3e71d7d9e11d24b93c2ec6d6cbf4d25e1be9c2