more flexible reified check for final classes
Summary:
## What

loosen a reified check, see test case

## Why

The reified check (among other things) prevents trying to reify an unknown type. So it is understandably careful with late static bindings, since there are cases where the type cannot be known statically.

However, in final classes, `this` isn't really late static binding: we know that it just refers to the lexically enclosing class and so the type is known.

Reviewed By: andrewjkennedy

Differential Revision: D75142486

fbshipit-source-id: 2541472b4a9d9431989f9dea41477c612bb7e26b