Fix a bug in how ScopedServerInterfaceThread creates clients
Summary:
TLRD: Before this change trying to use ScopedServerInterfaceThread::newClient() to create client using binary protocol would crash the process.

ScopedServerInterfaceThread::newChannel takes executor, channel
functor, numThreads and protocol. The thing is ScopedServerInterfaceThread::newClient calls are passing protocol in place of numThreads.

Because protocol is just an enum (not enum class), it is implicitly converted into int and default protocol T_BINARY_PROTOCOL is used instead of T_COMPACT_PROTOCOL that is default for ScopedServerInterfaceThread::newClient.

This worked only because T_COMPACT_PROTOCOL is when implicitly converted to int results in value of 2 for numThreads. But, if one tries to actually create client with T_BINARY_PROTOCOL who's int value is 0, the process crashes with devision by 0.

Reviewed By: owwlo

Differential Revision: D75345838

fbshipit-source-id: 19253c61a22299e0a2ff44de6b49a11ec620d843