Allow converting individual nodes to TypeSystem
Summary:
Avoids the need to roundtrip from a definition node to its URI to convert between systems, which also adds support for nodes without URIs.
Accessing the URI of a node that does not have one is changed to throw, which will ensure we don't accidentally serialize it as part of the type system, use it to create an Any, or otherwise leak the partial nodes.

Reviewed By: pranavtbhat

Differential Revision: D75178791

fbshipit-source-id: c09b43ec6c2a190c5749fb2c22e866e51f152c0b