Remove bogus use of object type in UNSAFE_CAST
Summary:
The `object` type no longer exists except as a temporary definition in `fbcode/hphp/runtime/ext/core/hhi/object.hhi` where it is bound to `dynamic`. Certainly it makes no sense in an `UNSAFE_CAST`.

Also fix up the test that triggers the lie detector so that it type-checks under `hh`.

Reviewed By: viratyosin

Differential Revision: D75528761

fbshipit-source-id: 2fdfa99976ee35ab9de851003533bc34ebbade15