bind ServiceInterceptor
Summary:
I was just trying to make a python test that could verify that service interceptors were being invoked from cpp, but then I realized this approach could be viable for allowing users to author service interceptors in Python.

The basic idea is:
- The Python user adds their custom Python-language implementations to a class that derives `AbstractServiceInterceptor` interface.
- `PyObservableServiceInterceptor` serves as a bridge between python implementations in user impl and C++.
  - The `getattr` in ctor lets us skip this step for subsequent calls (Cython getattr are pre-compiled).
  - It also specializes the template so that both `ConnectState` and `RequestState` are `PyObject*`.
  - Whatever python `object` is returned by `onConnect` is passed to the other interceptors, and whatever python `object` is returned by `onRequest` is passed to `onResponse`.

`RequestInfo`, `ResponseInfo`, `ConnectionInfo`, are lazy wrappers around the corresponding C++ concepts. Lazy here means they do minimum of work, only copying pointers. Any copies / decodes are deferred until corresponding property is accessed. Only the subset of fields that seem immediately interesting to Python are bound for now; others can be extended later.

The current state is that `RocketServerHandler` already invokes `onConnection` and `onConnectionClosed` regardless of language. The main remaining task is invoke `onRequest` and `onResponse` from `PythonAsyncProcessor`. This is required both for C++ implemented interceptors and Python ones like in the test.

If we want to "productionize" this for non-test use cases with Python implementations, the following are needed:
- translating args / response (doable if the cpp is populated with the serialized IOBuf)
- async-ify onRequest / onResponse (they have async C++ interface, unlike the connection interceptors).

Reviewed By: prakashgayasen

Differential Revision: D75273371

fbshipit-source-id: 65d9b021b7b3035e51f40b1aaed50135b5e5e6f4