Avoid intermediate centroids buffer in TDigest merge
Summary:
`TDigest::merge()` currently accumulates all the digests into an intermediate array of centroids, then sort it by using a recursive in-place merge (since each digest is already sorted). This requires the allocation of a large array, copying various kBs of data, and a relatively slow in-place merge.

Instead, we can use the classical textbook solution of merging the centroids in a streaming fashion using a heap, which is both faster and simpler.

Reviewed By: DenisYaroshevskiy

Differential Revision: D75216993

fbshipit-source-id: 771e1dc824da615c3c402433c2350e71e44a0128