Optimize final sort in TDigest::merge
Summary:
The process of merging sets of sorted centroids should result in a sorted list of centroids, but numerical error can produce perturbations in the order, so the current implementation does a fix-up `std::sort` at the end.

This is unnecessarily expensive (for example, libstdc++'s `std::sort` does not special-case already-sorted arrays): since the result is already sorted virtually all the time, and if a perturbation does happen it should be small, we can do insertion sort as we go, which in the overwhelmingly common case just costs one comparison.

Reviewed By: DenisYaroshevskiy

Differential Revision: D75217947

fbshipit-source-id: 36a3bd0e41c9eac2ad9acba294c7b0c57d447a67