Capture back-references using "x"_id = ... tags
Summary:
For example, this lets a scope take a ref to a closure-owned result collector -- or, several background tasks use the same closure-owned result collector.

The only alternative to backreferences is nesting multiple closures, which is both uglier (due to extra nesting) and less efficient (because coro frames can be hard to optimize away).

Reviewed By: ispeters

Differential Revision: D75415806

fbshipit-source-id: be32f7fa82903b66231385f86891b72e76a27074