Work around toolchain limitation
Summary:
Thrift is being built on a toolchain where instantiating `std::pair` on an incomplete type will fail to compile.
F14FastMap is an alias that resolves to either the value or vector flavors depending on the size of `std::pair<K,V>`.
Evaluate that condition manually and inline the result to avoid the problematic instantiation.

Reviewed By: praihan

Differential Revision: D75617919

fbshipit-source-id: f4fc251a5bbb80d0611c8abbdb10f09ef8d22b15