Add missing cases default statements
Summary:
This diff is part of a stack that allows us to turn on stricter compiler flags
that prevent clang from compiling code with a switch-case that should be
illegal.  In particular, consider the following code
```
enum class E {
  ONE = 1,
  TWO = 2
}
int foo(E e) {
  switch (e) {
    case E::ONE: {
      return 1;
    }
    case E::TWO: {
      return 2;
    }
  }
}
```

- Calling foo(static_cast<E>(3)) compiles in clang (whereas it does not in gcc).
  This caused S519734. -Wswitch-default prevents us from hitting this UB.
- If we do not enable -Wswitch-enum, we do not get exhaustiveness checking for
  enum-based switch-case statements.  This diff enables that flag as well.

This diff fixes a similar issue in FieldMaskUtil and visit_union() (along with
all its usecases) that can cause UB today because of this.

Reviewed By: iahs

Differential Revision: D75120135

fbshipit-source-id: 3d12935e0b22e4416d214430db5cc45288016713