Make sweeping HSLFileDescriptors a no-op
A trivial CLI program that outputs to stdout, such as this one:
```hack
use namespace HH\Lib\IO;

<<__EntryPoint>>
async function main(): Awaitable<void> {
//    var_dump(new Test());
  await IO\request_output()->writeAllAsync("Hello World!\n");
}
```
will abort on exit on a dev build of HHVM with the following assertion
failure:
```
hphp/runtime/base/countable.h:285: bool HPHP::MaybeCountable::countedDecRefAndCheck(): assertion `!tl_sweeping' failed.
```

GDB points to HSLFileDescriptor as the culprit:
```
(gdb) bt
    msg=...) at /hhvm/hphp/util/assertions.cpp:73
```

HSLFileDescriptor may either be initialized with a raw FD or a req::ptr
holding a File. In the former case, sweeping is already a no-op since
the extension maintains a request-local set of FDs to be closed and
processes it on request shutdown. For the File case, we only ever seem
to pass in builtin files (namely STDOUT and co.), which should likewise
already be closed by BuiltinFile::sweep() /
BuiltinFiles::requestShutdown().

So, define a no-op sweep() for HSLFileDescriptor to avoid triggering
redundant cleanup and the assert.