pass watchman_sockname option through to ClientCheckStatus
Summary:
As of D75139664, the `Edenfs_watcher` test suite in `hack/test/integration/test_edenfs_file_watcher.py` spawns its own Watchman instance instead of using the "global" system instance.
To this end, the tests set the `watchman_sockname` option in `hh.conf` to the socket of the Watchman instance used for testing.

Prior to this diff, the `watchman_sockname` option is respected whenever we talk to Watchman in Hack, expect for `ClientCheckStatus`.
In `ClientCheckStatus.watchman_get_raw_updates_since`, we were previously calling `Watchman_lwt.watch_project ~root ~sockname:None` before actually sending a query to Watchman. This means that we first register our root folder with the global (!) Watchman instance (if it wasn't registered already) before sending it a query later in `watchman_get_raw_updates_since`.

As a result, if `watchman_sockname` is set, the server may talk to one Watchman instance, and then the client talks to to another one ( = the global one), but at least it would register the folder with the global Watchman instance first. I experimented a bit with this and this crossing of instances seems to just work, but for consistency I would like to make sure that everything just talks to the single Watchman instance identified by `watchman_sockname`.

Therefore, this diff pipes the `watchman_sockname` option from `ServerLocalConfig` to `ClientCheckStatus.watchman_get_raw_updates_since` and makes sure we use it when talking to Watchman.

Differential Revision: D75140228

fbshipit-source-id: 881520af62be9373c154c627d6d1d16af648fea7