Use cast to dynamic instead of resource in enum class label implementation
Summary: The Hack lie detector tripped on a cast to `resource` inside the implementation of `valueOf` and `nameOf` for enum class labels. Resources are a legacy representation, used only when `hhvm.emit_native_enum_class_labels=false`. We should kill the legacy implementation; pending that, use `dynamic` instead of `resource` in the cast.

Reviewed By: vassilmladenov

Differential Revision: D75594423

fbshipit-source-id: 02b08b6f4875dfc0240d88d28e304eed8984862e