Make type_structure raise implicit string to class conversion
Summary:
I am aiming to replace occurrences of
```
$x = type_structure($c, 'TFoo')['classname'];
new $x();
```
with
```
public static function getTFoo(): class<this::TFoo> {
  return HH\ReifiedGenerics\get_class_from_type<this::TFoo>();
}

...
$x = $c::getTFoo(); // log if $c is a string
new $x();
```
and the latter will log an implicit string to class conversion if `$c` is a string. This change makes it so we do equivalent logging in `type_structure(...)` calls, considering them in the set of expressions that do implicit conversion.

In other words, if a class is loaded from a string for type_structure, we want to make sure it will get marked as `<<__DynamicallyReferenced>>`.

# HHBBC

Breaks the optimization for static strings that eliminates the expression, falling back on the runtime to raise the notice. If this ends up being too expensive, my fallback plan is to add a bytecode similar to `RaiseClassStringConversionNotice` and emit it when we have BStr.

Reviewed By: jano

Differential Revision: D73490011

fbshipit-source-id: bc4bb75602eb11c1bf8215d21410a890e6be6bde