fix mock-ability of clients
Summary:
My initial theory for fixing this was wrong. We can't directly map `Py3ClientName = PythonClientName.Async` because too many interfaces on the thrift-python side expect the input to be the `SyncOrAsync` wrapper class.

So to support mocks, we can just add stub methods in auto-migrate. But until we see evidence these are needed for non-mock situation, let's leave them as `raise RuntimeError`. It should be impossible to successfully call `__aenter__` in a non-mock situation because the `_connect_future` won't  complete; it's normally set by a callback in `get_client`.

Reviewed By: annakukliansky, Sunnie912

Differential Revision: D75822268

fbshipit-source-id: 73ad6f354d61855549fab950c8aafd575c7efcb2