imlement Edenfs_watcher.get_changes_sync
Summary:
This diff implements `Edenfs_watcher.get_changes_sync`, which in turn powers `ServerNotifier.get_changes_sync` if `edenfs_file_watcher_enabled` is set.

In contrast to its `_async` counterpart, this version will not just return any changes that the worker has picked up so far, but directly ask Eden if any further changes have happened since then.

The main issue is the coordination between the worker and what happens in `get_changes_sync`: In the latter, we a) want to get whatever changes the worker has picked up and b) then check with Eden if any further changes happened since then. We need to make sure that whatever changes we pick up in b) are not processed again by the worker in a way that they would be returned to the callers twice.

From what I can tell, calls to `ServerNotifier.get_changes_sync` are much more rare than those to `get_changes_async`; they are only triggered by certain client commands (instead of being triggered in a hot loop or something like that).

Because of that, I decided to go with the simplest coordination strategy: We simply stop the worker, get the changes it processed until then, and then start a new worker at the end of `get_changes_sync` which picks things up from there.

Differential Revision: D75872304

fbshipit-source-id: e0f53c9224f15460256f0a444d31328056c8722d