Support fixpoint class refinements
Summary:
Where we have a fixpoint class refinement, e.g.:

```
interface I {
  abstract const type TNext as I with { type TNext = this::TNext }
}
```

localization will call into type constant expansion which itself will make a recursive call to localize the refined definition of detect `TNext = this::TNext`. At this point, type constant expansion will determine there is a cycle resulting in a `TAny` in place of the refined `this::TNext`.

This diff adds a substitution when we are projecting from an expression dependent type (our representation of implicit, existential tags) so that the definition is replace with the mangled generic we currently use to represent a type constant projected off an existential tag.

Reviewed By: andrewjkennedy

Differential Revision: D75819777

fbshipit-source-id: 39cb7d69671a82415209a6e6e6d80ce906e065ca