generated is_auto_migrated global
Summary:
This ended up being weirdly complicated. The issue is that we need to access auto-migrate `bool` from `lib/py3/types.pyx`. But we can't depend on any generated `-py3-types` without creating a circular dependency. So our options are as follows:
- generate a bool in its own target `python_library`
- use `-D` compiler option and expose it via cython.

The problem with `-D` constant approach is we're forced to generate two separate versions of `//thrift/lib/py3:types`, one for auto and one for default, then make `py3:types` an alias to the proper one based on auto-migrate. This adds too much complexity and more risk of issues.

Instead, opt to generate `types_auto_migrated.py`, only containing a bool. We don't need to do this in every thrift file; we can limit it to only thrift files in `thrift/lib`.

Reviewed By: createdbysk, prakashgayasen

Differential Revision: D75891062

fbshipit-source-id: 8fe7b3e64f640065d2e284e3497c851ad6d4b6fe