make isinstance Py3 containers pass
Summary:
Using the global from previous diff, we can now make `isinstance(obj, py3.types.List)` continue to return `True` even when auto-migrate enabled. This uses the metaclass `__instancecheck__` override trick.

This was trickier than for `Struct` or `Union` because thrift-python containers are not generated, so we don't have a chance to easily insert the is auto-migrated flag.

Reviewed By: prakashgayasen, createdbysk

Differential Revision: D75898465

fbshipit-source-id: f1c100764bf25d3568e005534b8d28409979d295