Auto-fix missing semicolon on last line of a cell
Summary:
## What

When converting a Hack notebook from .ipynb (JSON) to Hack (.php), try to help the user if they forgot a semicolon at the end of the cell.

See test case for an example.

## Details

Specifically, we add a semicolon to a cell containing top-level statements if-and-only-if doing so would bring the parse error count from 1 to 0.

## Why

Users routinely omit the final semicolon in a cell consisting of top-level statements.

Users probably omit the semicolon because the HHVM debugger allows it and notebooks can run successfully even without the final semicolon (since they are built on the debugger).

This change makes the conversion script more tolerant, saving users time and confusion.

At the same time, we want the conversion to be predictable and maintainable: so we are very unfancy here. If there are other syntax errors, we leave the code as-is and let the user use CI and language server etc. to fix the errors.

Reviewed By: lillyfwang

Differential Revision: D75902802

fbshipit-source-id: fc5b7f71fc817451cb5e525fae0a4a0fabe05f06