Fix race between onThreadExit and destroy of ThreadLocal objects.
Summary:
The 'destroy' call for ThreadLocal objects was doing racy access to
ThreadEntry objects which could result in segfaults or other bad access when
racing with an exiting thread. It was possible for destry to access an already
freed ThreadEntry and for both threads to concurrenly work on the same element.

Reviewed By: IosifSpulber

Differential Revision: D75743184

fbshipit-source-id: 926fe07ba7a11152e22d68011fa0499f678e15b3