AsyncSocket: allow returning exception from ReadAncillaryDataCallback::ancillaryData
Summary:
**Context for this stack:**

Using `AsyncFdSocket` to connect to a local thrift server that uses service framework over UDS and passing header fds currently does not work. This is because the server logic depends on `PeekingAcceptorHandshakeHelper` that peaks into the socket data and if it doesn't look like TLS data forwards to `UnencryptedAcceptorHandshakeHelper`. `UnencryptedAcceptorHandshakeHelper` does not use `AsyncFdSocket` so it misses any cmsg fds. Changing `UnencryptedAcceptorHandshakeHelper` to use `AsyncFdSocket` alone is not sufficient as the initial peeking operates on `AsyncSSLSocket` and can also miss cmsg fds.

To fix this, this stack moves the queue of cmsg fds from `AsyncFdSocket` to the base `AsyncSocket` class, and sets `readAncillaryDataCallback` in `AsyncSSLSocket` to allow receiving cmsg fds when the `AsyncSSLSocket` is in the unencrypted state.

**This diff:**
Change `ReadAncillaryDataCallback::ancillaryData` return type to allow returning and exception. This is required to untangle the callback from `AsyncFdSocket` and use it in `AsyncSSLSocket` later in the stack.

Reviewed By: snarkmaster

Differential Revision: D75676139

fbshipit-source-id: 45eab8a57035fb7ffbf1f91391ccad7332b0b5ac