Enforce STRICT_CHECK in http_parser
Summary:
For T225434539.

Previously, STRICT_CHECK was a no-op, but we enforce STRICT_CHECK here and rename the no-op "strict" check to NOOP_CHECK. We use the new STRICT_CHECK in the case `s_chunk_size_almost_done` so that after a carriage return, a character that isn't a line feed is rejected.

Reviewed By: hanidamlaj

Differential Revision: D76075261

fbshipit-source-id: bebc189285f41845ac3f7792240ab07ddd5c4df4