modify CallFuncEntry to take in callee as an SSATmp
Summary:
This diff changes CallFuncEntry to take in the callee as an SSATmp, and adds support for calling the func entry from a dynamic func object. The main change is that we now take in a `calleePrototype` to derive compile-time data from, but allow calling the func entry of any callable that has the same signature as the `calleePrototype`, with some limitations.

We now initialize the `m_funcEntry` field of Funcs to a new stub called `handleTranslateVirtualFuncEntry`, which calls `handleTranslateFuncEntry` with numArgs = the number of parameters inferred from the live function data. This is similar to what `callphpfe` does, which sets `numArgs` and calls `handleTranslateFuncEntry` for its own stubs. We don't want to set `numArgs` here, because we don't have smashable calls and would be incurring an extra instruction on the hot path.

While testing my change, I noticed that we weren't setting `m_funcEntry` for profiling translations in most cases. It turns out that the `newTranslation` code (which smashes call addresses etc) wasn't doing this since prologues & func entries are set later for optimized translations. Change it to do so for profiling translations.

A major limitation of the current approach is that we only support calling func entry when there are no default arguments, consistent with how we set `m_funcEntry` in [existing code](https://www.internalfb.com/code/fbsource/[20db558f1ed6d87544d20ac69d186b14a9443be4]/fbcode/hphp/runtime/vm/jit/service-request-handlers.cpp?lines=215-220). We talked about a couple ways to relax this restriction moving forward, but will focus on first moving prologues of virtual calls to the caller side first in what we expect to be the common case.

Reviewed By: jano

Differential Revision: D74002339

fbshipit-source-id: 82ed9e341bd89305b36bf975dd853015e0ff42af