simplify `svcreq::getFuncEntry` to unconditionally return the main func entry address
Summary:
We had complex logic on `svcreq::getFuncEntry` to deal with the complexity of passing in varying numbers of
arguments that didn't necessarily match the main entry point (when numArgs == numNonVariadicParams), and the
existing code had complexity to avoid registering the func entry in those cases.
We don't need that complexity as there's a single caller of `svcreq::getFuncEntry` where we know the arg count.

This diff simplifies the whole situation by introducing a new main func entry translation stub that translates
assuming `numArgs == numNonVariadicParams`, and updates the funcEntry field of the invoked function. This allows
us to remove the various checks in `svcreq::getFuncEntry`, and things are now morally much closer to simply fetching
the right funcEntry offset.

One related change: We want to avoid trying to jit things over and over when we run out of TC space, so some care is taken in the func entry translation stub to fall back to a noTranslate variant on persistent failures, so when we hit `setCachedForProcessFail`, we change the func entry of the failed func to a noTranslate stub.

Reviewed By: jano

Differential Revision: D75964653

fbshipit-source-id: 52b3d8de691fdecec706266619823c03c4e45527