Change How We Virtualize Operation Typing in Expression Trees
Summary:
# Problem
Expression trees today virtualize operations on values to calls on instance methods. For instance given the following:
```
ExampleDsl`1 === 2`;
```
This will virtualize for typing to roughly:
```
ExampleDsl::intType()->__tripleEquals(ExampleDsl::intType())
```

This way of virtualizing restricts operations to being performed on objects within expression trees. This shows up most prominently if the null literal wants to be treated as the type `null` and compare it to another value. i.e.
```
ExampleDsl`null === {$nullable}`
```
If `ExampleDsl::nullType() : null` then the current virtualization scheme won't work.

# Solution
We can increase the flexibility of our virtualization scheme by adding an additional layer of indirection. Instead of virtualizing directly to an instance method call, we wrap the first operand with a call to `Dsl::operationType($operand)`. This gives an opportunity to route a non-object type to an object type that defines the operation methods.

This added flexibility comes at minimum cost. If the excessing behavior is desired then the operationType method can be implemented as the identity function and the behavior will remain the same.

As a worked out example given:
```
ExampleDsl`1 === 2`;
```
Before:
```
ExampleDsl::intType()->__tripleEquals(ExampleDsl::intType())
```
After:
```
ExampleDsl::operationType(ExampleDsl::intType())->__tripleEquals(ExampleDsl::intType())
```

# Implementation
We apply this transform in three cases:
1. For binary expressions of the form `operand1 op operand2`
2. For unary expressions of the form `op operand`
3. For implicit casts to bool which is abstracted into the `boolify` function

Given the similarities with binary and unary expressions, I introduced a function `virtualize_operation` that both can use.

The remainder of the diff is mainly updating relevant tests. This usually involved adding the operationType method as the identity function and updating generated tests.

Reviewed By: viratyosin

Differential Revision: D76289896

fbshipit-source-id: e246aecea6c508b406d96371db38beee25936834