Add new annotation scope for function thrown exceptions
Summary:
- Add a new annotation scope for RPC function thrown exception declarations
- Add thrown exception scope to `rust.Name`

This diff simply adds the scope, so it can be synced to www. A follow-up diff in the stack will validate and emit a compiler error.

Adding the scope and compiler error in one-shot fails in CI, as the new scope is not yet present in WWW's synced `scope.thrift`.

Reviewed By: iahs

Differential Revision: D76289290

fbshipit-source-id: 1c11630aca489c36535182c2a1ee81d6be850d84