Add error for recursive dependencies in where clauses of case types
Summary:
Recursion via where clauses like those in the test cases added in this diff can result in infinite recursions whenever we try to add the clause as a constraint or check the clause in subtyping.

This diff takes the approach of preemptively banning any recursive mention at all. A more precise restriction (https://www.cis.upenn.edu/~bcpierce/papers/variance.pdf) is a followup task (T225633579) (optional pending an actual need for a recursive mention)

The way I've implemented this is to have a recursive visitor that can find these recursive mentions. I use this to both (a) detect and report the error and (b) exclude errant where clauses. I'm not super happy about having the callsites just have to know they need to check for recursive mentions, but not sure if there's a better way.

Reviewed By: dlreeves

Differential Revision: D75460599

fbshipit-source-id: f3bd7c850e470c3341bf88f98e89ebb693fee7ab