Avoid flattening the same constant multiple times in HHBBC
Summary:
We may copy in abstract-with-default constants to a class's constant list before
trait flattening occurs. This causes us to have duplicate constants in the  constants list if the flattened trait contains the same constant name. This breaks the desired invariant of one constant-per-name and causes lots of downstream issues.

To remedy this, remove the extra constant from the "trait cns" list of the relevant class.

The newly added assert reveals that we have pre-existing copies of type and context constants in `cls.constants` (see the comments in `add_constant`). Make the flatten_traits override behavior explicit by searching and updating those constants as appropriate.

Reviewed By: paulbiss

Differential Revision: D76129284

fbshipit-source-id: 0c7eacda839dcdec8f46ba684d73f438281fe271