Do not attempt to drive the Executor when __del__ after a fork
Summary:
In python 3.12 it seems eventloops are destroyed after a fork.   See: https://www.internalfb.com/phabricator/paste/view/P1836944088?lines=34

In the case where the PID has changed do not attempt to drive the executor inside the delete method.

This was discovered in S529528, because of a preexec_fn being used in a subprocess.run causing python to execute code between the fork and the exec.

Reviewed By: aleivag

Differential Revision: D76365095

fbshipit-source-id: d868e95a9651ff176f34b0b3bfd2433e7aab0776