Add debugger command and stdout hook for streaming.
Summary:
## Context
Fbdbg is an unowned and unmaintained Hack debugger that's used in Hack Bento Notebooks and on the command line as a repl.

Although fbdbg enables debugging capabilities, it's rarely used among Hack programmers, with the VSCode debugging extension being preferred.

Deprecating fbdbg allows us to
1. Reduce support burden on the HHVM team
2. Enable more capabilities for Hack notebooks easily

["Deprecating fbdbg" Google Doc](https://docs.google.com/document/d/12CyH78_LT6CbXMQF603-_75DwkVlpmYo2MW9bfEFMM4/edit?tab=t.0)
## This diff
This diff creates a new command called `CmdEvalStream` which evaluates user commands in the same way as `CmdEval` and streams outputs to the client.

Whenever HHVM calls `DebuggerStreamStdoutHook`, the outputs are streamed to the client instead of just being appended to a string buffer like in `DebuggerStdoutHook`.

Reviewed By: nt591

Differential Revision: D75472812

fbshipit-source-id: efb3f83789bf9c90b869fb108bd4e5fa4983e096