Edenfs_watcher: handle directory renames
Summary:
When a folder gets renamed, Eden gives us a `LargeChangeNotification::DirectoryRenamed` notification, containing the old an new path.

In contrast, hh_server expects to be told about all the affected files individually. To bridge this gap, we handle being informed about `foo` being renamed to `bar` as follows:
1. Look up all the files in folder `bar`. Notify `hh_server` about all of them.
2. For each file in step 1, determine where it must have been beforehand (by translating the path `foo` to `bar`, effectively inverting the renaming. Notify hh_server about these paths, so that it can note that the file no longer exists at this path.

## Facebook

This addresses T226746776 in the sense that it fixes the problem described there, but only when running with `edenfs_file_watcher_enabled` set.

Differential Revision: D76124936

fbshipit-source-id: 91f39a8588ef9041efb8f7f381b5f930c985665b