Introduce a new flag PNRc to the IR opcode spec
Summary:
This will help us determine when we are creating absolutely new values with a refcount of 1. This information will help us set the upper bound on the refcount of destination asets in refcount-opts.

I added this flag to those instructions that seemed like an obvious case. I will add this flag to other not-so-obvious instructions as well in the future.

Reviewed By: jano

Differential Revision: D75106620

fbshipit-source-id: 683d39a307e169a60c6616582ae0d5b93ca43294