Converge early if propagating RCState across a cycle's back edge
Summary:
I noticed that for upper bound analysis, it would propagate new RCState across a back edge. During propagation, it would increase the upper bound on some Aset in the destination block's RCState by 1. Due to this change, we reprocess that destination block again. We reach the block with the back edge and the cycle continues until we converge the upper bound of that Aset to INT_MAX. We end up processing the cycle INT_MAX times and it was timing out during the tests.

The lower bound also has the same issue but we prob don't see it in production because for lower bound, it converges to 0. My guess is that in production and our tests, the lower bound of any Aset is a small number. As long as it's less than 1e5, it won't be noticeable.

Reviewed By: jano

Differential Revision: D75771872

fbshipit-source-id: 4b35cb952830f4283e1e6567fc7d01e5ba1506d0