Add a runtime check that a non-async method cannot override an async method
Summary:
Currently, we allow an async method to be overridden by a non-async method. For non-async methods that return an un-awaited awaitable, we only enforce that the return value is an Awaitable object. If such a method overrides an async method, it is possible to construct method hierarchies where the return type of the overriding method is wider than its parent.

This check already exists in the type checker. Add the check to the runtime as well.

Reviewed By: jano

Differential Revision: D76155725

fbshipit-source-id: 4a032ccab4ef4a1ba86604a707072d00533f914f