Do not post waiters while holding the state spinlock in fibers TimedMutex
Summary:
Fiber `TimedMutex` implementations currently wake the readers while holding the spinlock. Waking up can be a very expensive operation (potentially involving syscalls) and it can result in a preemption, so they should not be done in the critical section.

This is generally easy except for one case: when the waiting thread times out, the waiter may be removed from the list and destroyed, and since the waker is not holding the spinlock anymore there is no protection.

So, we change the waiter to delay destruction until it has been awoken. This creates the simple contract that whoever removes the waiter from the list has to awake it, which guarantees that the waiter is never destroyed in a non-awoken state.

Reviewed By: ilvokhin

Differential Revision: D76337625

fbshipit-source-id: d097c211e3b16a669076a932ff407677dda96b1d