Remove inline bound from Tnewtype
Summary:
The `Tnewtype` constructor for `newtype` uses contains inline in the type its bound i.e. the `as` type from the definition instantiated with its type arguments. This contrasts with with practice elsewhere (bounds on abstract type constants, bounds on generics, inheritance for classes, etc) where we compute the bounds on demand. When creating a `Tnewtype` we must be careful that the bound actually matches the definition, and in general it's a potential source of bugs. Also, it doesn't easily extend to multiple bounds.

Instead, we now look up the bound on demand, using the `get_newtype_super` helper function. In a future diff, we will extend the "type-directed" interpretation of `TypeStructure<t>` (currently limited to direct shape indexing) by simply returning a bound from `get_newtype_super` that is determined by `t` .

A number of error messages get slightly disturbed, but otherwise this is a meaning-preserving change.

Reviewed By: vassilmladenov

Differential Revision: D71197797

fbshipit-source-id: e034aaf3121c449af7d188eedc67dc1b7f1069fe